Move Network test helper from util crate into text crate
This way, `util` does not depend on `clock`.
